python/pyproject.toml
[project]
name = "langchain-examples"
version = "0.1.0"
description = "LangChain v1.x 示例集合"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.10",
    "openai>=1.0.0",
    "chromadb>=0.4.0",
    "faiss-cpu>=1.7.0",
    "requests>=2.25.0",
    "beautifulsoup4>=4.9.0",
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "jupyterlab>=3.0.0",
    "ipykernel>=6.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "matplotlib>=3.5.0",
    "plotly>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "ruff>=0.1.0",
    "pytest>=7.0.0",
    "httpx>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["00-env", "01-hello-chain", "02-prompt-template", "03-memory-chat", "04-rag-qa", "05-agent-weather", "06-api-deployment", "07-advanced-agents", "08-structured-output", "09-multi-agent", "10-streaming-chat", "11-production-tracing"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]

[dependency-groups]
dev = [
    "black>=23.0.0",
    "ruff>=0.1.0",
    "pytest>=7.0.0",
    "httpx>=0.24.0",
]
